Fix sub filter
When you subtract a higher byte value from a lower byte value you get a
negative number. When you put that negative number in a binary and pull
it out with integer and the default signedness, it will be 256 plus that
negative number (e.g. -15 will be 241). If you add the higher byte back
and modulo 256 you get the original number back.