Create render helper to render multiple images
So, I couldn't figure out what was going wrong in erl_freetype but I
knew that only some of the binary passed into erl_freetype is actually
used. I need to know what the width of the bitmap was that was copied
from erl_freetype. I created a function to create up to 24 PNG files
with varying widths to figure out which image looked right.

The z.html web page was original to just display the z.png file but now
it displays all 24 PNG files that are generated by render.erl.

Fixed the scanline generation for writing PNG files. Before it was just
using all the pixels. ... Maybe I intended #png.pixels to be an list of
scanlines. Hmmm. That's probably it.

Use the include file instead of re-declaring all the records.